+++ src/main/java/com/graduationproject/studentinformationsystem/common/model/dto/response/SisBasePersonalInfoResponse.java
@@ -1,6 +1,5 @@
-package com.graduationproject.studentinformationsystem.common.util.controller.response;
+package com.graduationproject.studentinformationsystem.common.model.dto.response;
 
-import com.graduationproject.studentinformationsystem.common.model.dto.response.SisBaseResponse;
 import lombok.Getter;
 import lombok.experimental.SuperBuilder;
