+++ src/main/java/com/graduationproject/studentinformationsystem/university/officer/model/dto/response/OfficerPersonalInfoResponse.java
@@ -1,6 +1,6 @@
 package com.graduationproject.studentinformationsystem.university.officer.model.dto.response;
 
-import com.graduationproject.studentinformationsystem.common.util.controller.response.SisBasePersonalInfoResponse;
+import com.graduationproject.studentinformationsystem.common.model.dto.response.SisBasePersonalInfoResponse;
 import lombok.Getter;
 import lombok.experimental.SuperBuilder;
