+++ src/main/java/com/graduationproject/studentinformationsystem/university/student/model/dto/response/StudentAcademicInfoResponse.java
@@ -1,6 +1,6 @@
 package com.graduationproject.studentinformationsystem.university.student.model.dto.response;
 
-import com.graduationproject.studentinformationsystem.common.util.controller.response.SisBaseAcademicInfoResponse;
+import com.graduationproject.studentinformationsystem.common.model.dto.response.SisBaseAcademicInfoResponse;
 import com.graduationproject.studentinformationsystem.university.department.model.dto.response.DepartmentResponse;
 import com.graduationproject.studentinformationsystem.university.student.model.enums.StudentClassLevel;
 import com.graduationproject.studentinformationsystem.university.student.model.enums.StudentDegree;
